# Replace `node-sass` package with `sass`