--- a/package.json
+++ b/package.json
@@ -436,7 +436,6 @@
     "exit-hook": "^2.2.0",
     "faker": "1.1.0",
     "fetch-mock": "^7.3.9",
-    "fibers": "^4.0.2",
     "geckodriver": "^1.19.0",
     "getopts": "^2.2.4",
     "grunt": "1.0.4",