--- a/packages/kbn-optimizer/package.json
+++ b/packages/kbn-optimizer/package.json
@@ -29,6 +29,7 @@
     "css-loader": "^3.4.2",
     "del": "^5.1.0",
     "execa": "^4.0.2",
+    "fibers": "^4.0.2",
     "file-loader": "^4.2.0",
     "istanbul-instrumenter-loader": "^3.0.1",
     "jest-diff": "^25.1.0",