--- a/packages/kbn-optimizer/src/worker/webpack.config.ts
+++ b/packages/kbn-optimizer/src/worker/webpack.config.ts
@@ -269,6 +269,7 @@ export function getWebpackConfig(bundle: Bundle, worker: WorkerConfig) {
 
                           return { file: url };
                         },
+                        fiber: require('fibers'),
                       };
                     },
                   },