--- a/packages/kbn-ui-framework/package.json
+++ b/packages/kbn-ui-framework/package.json
@@ -41,6 +41,7 @@
     "core-js": "^3.6.4",
     "css-loader": "^3.4.2",
     "expose-loader": "^0.7.5",
+    "fibers": "^4.0.2",
     "file-loader": "^4.2.0",
     "grunt": "1.0.4",
     "grunt-babel": "^8.0.0",