--- a/src/legacy/server/sass/build.test.js
+++ b/src/legacy/server/sass/build.test.js
@@ -49,12 +49,7 @@ it('builds light themed SASS', async () => {
 
   expect(readFileSync(targetPath, 'utf8').replace(/(\/\*# sourceMappingURL=).*( \*\/)/, '$1...$2'))
     .toMatchInlineSnapshot(`
-    "foo bar {
-      display: -webkit-box;
-      display: -webkit-flex;
-      display: -ms-flexbox;
-      display: flex;
-      background: #e6f0f8 url(./images/img.png) url(ui/assets/favicons/favicon.ico); }
+    "foo bar{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;background:#e6f0f8 url(./images/img.png) url(ui/assets/favicons/favicon.ico)}
     /*# sourceMappingURL=... */"
   `);
 });
@@ -74,12 +69,7 @@ it('builds dark themed SASS', async () => {
 
   expect(readFileSync(targetPath, 'utf8').replace(/(\/\*# sourceMappingURL=).*( \*\/)/, '$1...$2'))
     .toMatchInlineSnapshot(`
-    "foo bar {
-      display: -webkit-box;
-      display: -webkit-flex;
-      display: -ms-flexbox;
-      display: flex;
-      background: #232635 url(./images/img.png) url(ui/assets/favicons/favicon.ico); }
+    "foo bar{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;background:#232635 url(./images/img.png) url(ui/assets/favicons/favicon.ico)}
     /*# sourceMappingURL=... */"
   `);
 });
@@ -103,12 +93,7 @@ it('rewrites url imports', async () => {
 
   expect(readFileSync(targetPath, 'utf8').replace(/(\/\*# sourceMappingURL=).*( \*\/)/, '$1...$2'))
     .toMatchInlineSnapshot(`
-    "foo bar {
-      display: -webkit-box;
-      display: -webkit-flex;
-      display: -ms-flexbox;
-      display: flex;
-      background: #232635 url(__REPLACE_WITH_PUBLIC_PATH__foo/bar/images/img.png) url(__REPLACE_WITH_PUBLIC_PATH__ui/favicons/favicon.ico); }
+    "foo bar{display:-webkit-box;display:-webkit-flex;display:-ms-flexbox;display:flex;background:#232635 url(__REPLACE_WITH_PUBLIC_PATH__foo/bar/images/img.png) url(__REPLACE_WITH_PUBLIC_PATH__ui/favicons/favicon.ico)}
     /*# sourceMappingURL=... */"
   `);