--- a/x-pack/plugins/canvas/shareable_runtime/webpack.config.js
+++ b/x-pack/plugins/canvas/shareable_runtime/webpack.config.js
@@ -193,7 +193,7 @@ module.exports = {
               },
               webpackImporter: false,
               sassOptions: {
-                outputStyle: 'nested',
+                outputStyle: 'compressed',
                 includePaths: [path.resolve(KIBANA_ROOT, 'node_modules')],
               },
             },