## Overview

In this PR, I changed all of our usage of [`node-sass`](https://www.npmjs.com/package/node-sass) to [`(dart-)sass`](https://www.npmjs.com/package/sass) instead. The latter is the reference implementation of sass; it is more well maintained, has lightweight dependencies, and doesn't rely on C bindings*.

### Performance

`node-sass` is generally faster than `sass` that is run as native JavaScript. However, based on some anecdotal manual testing, I didn't notice much of a performance difference for Kibana.

<details>
<summary>
Before changes
</summary>

```
np bld    log   [12:44:14.473] [success][@kbn/optimizer] 57 bundles compiled successfully after 99.8 sec, watching for changes
...
optmzr    log   [12:44:41.998] [info][optimize:dynamic_dll_plugin] Need to compile the client vendors dll
optmzr    log   [12:44:42.001] [info][optimize:dynamic_dll_plugin] Client vendors dll compilation started
optmzr    log   [12:45:05.281] [info][optimize:dynamic_dll_plugin] Client vendors dll compilation finished with success
optmzr    log   [12:45:17.549] [info][optimize:dynamic_dll_plugin] Finished all dynamic dll plugin tasks
optmzr    log   [12:45:17.551] [info][optimize] Optimization success in 120.33 seconds
optmzr    log   [12:45:22.639] [info][optimize:dynamic_dll_plugin] No need to compile client vendors dll
optmzr    log   [12:45:22.639] [info][optimize:dynamic_dll_plugin] Finished all dynamic dll plugin tasks
optmzr    log   [12:45:22.640] [info][optimize] Optimization success in 4.15 seconds
```

</details>

<details>
<summary>
After changes
</summary>

```
np bld    log   [12:50:32.665] [success][@kbn/optimizer] 57 bundles compiled successfully after 102.4 sec, watching for changes
...
optmzr    log   [12:51:19.476] [info][optimize:dynamic_dll_plugin] Need to compile the client vendors dll
optmzr    log   [12:51:19.493] [info][optimize:dynamic_dll_plugin] Client vendors dll compilation started
optmzr    log   [12:51:38.600] [info][optimize:dynamic_dll_plugin] Client vendors dll compilation finished with success
optmzr    log   [12:51:50.747] [info][optimize:dynamic_dll_plugin] Finished all dynamic dll plugin tasks
optmzr    log   [12:51:50.750] [info][optimize] Optimization success in 98.68 seconds
optmzr    log   [12:51:56.947] [info][optimize:dynamic_dll_plugin] No need to compile client vendors dll
optmzr    log   [12:51:56.948] [info][optimize:dynamic_dll_plugin] Finished all dynamic dll plugin tasks
optmzr    log   [12:51:56.948] [info][optimize] Optimization success in 5.17 seconds
```

</details>

\* Note: I did include `fibers` in a couple of consumers to speed up async rendering, but this comes with prebuilt binaries instead of relying on `node-gyp`. See https://github.com/elastic/kibana/pull/61722#discussion_r399683241.